Refactor: 마이페이지 관심분야 설정 API 경로를 mypage로 변경
